Move class that defines publisher and subscriber properties to pubsub module
